+++ tests/test_orm.py
@@ -11,51 +11,6 @@ from pyairtable.orm import fields as f
 from pyairtable.testing import fake_meta, fake_record
 
 
-def test_model_missing_meta():
-    """
-    Test that we throw an exception if Meta is missing.
-    """
-    with pytest.raises(AttributeError):
-
-        class Address(Model):
-            street = f.TextField("Street")
-
-
-def test_model_missing_meta_attribute():
-    """
-    Test that we throw an exception if Meta is missing a required attribute.
-    """
-    with pytest.raises(ValueError):
-
-        class Address(Model):
-            street = f.TextField("Street")
-
-            class Meta:
-                base_id = "required"
-                table_name = "required"
-                # api_key = "required"
-
-
-def test_model_empty_meta():
-    """
-    Test that we throw an exception when a required Meta attribute is None.
-    """
-    with pytest.raises(ValueError):
-
-        class Address(Model):
-            Meta = fake_meta(api_key=None)
-            street = f.TextField("Street")
-
-
-def test_model_overlapping():
-    # Should raise error because conflicts with .exists()
-    with pytest.raises(ValueError):
-
-        class Address(Model):
-            Meta = fake_meta()
-            exists = f.TextField("Exists")  # clases with Model.exists()
-
-
 class Address(Model):
     Meta = fake_meta(table_name="Address")
     street = f.TextField("Street")
@@ -73,13 +28,7 @@ class Contact(Model):
     created_at = f.CreatedTimeField("Created At")
 
 
-def test_repr():
-    record = fake_record()
-    assert repr(Contact.from_record(record)) == f"<Contact id='{record['id']}'>"
-    assert repr(Contact()) == "<unsaved Contact>"
-
-
-def test_model():
+def test_model_basics():
     contact = Contact(
         first_name="Gui",
         last_name="Talarico",
